refactor: address devenv review cleanup
